refactor(quiz): split logic into separate files
